requirements.txt
sphinx
pytest
pytest-pycodestyle
flask
testfixtures
pandas
scipy
matplotlib
pymodm
dnspython
Ipython
scikit-image
Pillow
